+++ src/components/kind/cards/small.js
