+++ src/components/symbol/cards/small.js
