+++ src/pageComponents/sidebar.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import {Route, Link, Switch} from 'react-router-dom'
 import {} from 'react-bootstrap'
-import RelatedKinds from '../kind/relatedKinds'
-import RelatedPantheons from '../pantheon/relatedPantheons'
-import RelatedSymbols from '../symbol/relatedSymbols'
+import RelatedKinds from '../components/kind/misc/related'
+import RelatedPantheons from '../components/pantheon/misc/related'
+import RelatedSymbols from '../components/symbol/misc/related'
 import SidePanel from './sidePanel'
 
 function Menu(props) {
@@ -41,7 +41,7 @@ function Menu(props) {
         </div> */}
 
         <div className="side-bar-section">
-            <div className="side-bar-panel">  
+            <div className="side-bar-panel">
                 <h5>Popular Dieties</h5>
                 <Link to='/symbol/Hecate'>Hecate</Link> (Greek)<br />
                 <Link to='/symbol/Horus'>Horus</Link> (Horus)<br />
@@ -53,24 +53,24 @@ function Menu(props) {
         <div className="side-bar-section">
             <div className="side-bar-panel">
                 <h5>Nature</h5>
-                <Link to='/collection/Plants and Herbs'>Plants & Herbs</Link> 
+                <Link to='/collection/Plants and Herbs'>Plants & Herbs</Link>
                 <Link to='/collection/Crystals'>Crystals</Link>
                 <Link to='/collection/Trees'>Trees</Link><br />
                 <h5>Knowledge</h5>
-                <Link to='/collection/People'>People</Link> 
+                <Link to='/collection/People'>People</Link>
                 <Link to='/collection/Stories'>Stories & Myths</Link>
                 <Link to='/collection/Teachings'>Teachings & Concepts</Link><br />
                 <h5>Divination</h5>
-                <Link to='/collection/Tarot Cards'>Tarot Cards</Link> 
+                <Link to='/collection/Tarot Cards'>Tarot Cards</Link>
                 <Link to='/collection/Runes'>Runes</Link>
                 <Link to='/collection/Divination Forms'>All</Link><br />
             </div>
         </div>
 
         <div className="side-bar-section">
-            <div className="side-bar-panel">  
+            <div className="side-bar-panel">
                 <h5>Beginner Categories</h5>
-                <Link to='/category/Magic 101'>Magic 101</Link> 
+                <Link to='/category/Magic 101'>Magic 101</Link>
                 <Link to='/category/Wicca 102'>Wicca 102</Link>
                 <Link to='/category/Occult 102'>Occult 102</Link><br />
             </div>
