+++ src/views/component/pantheon/page.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Symbols from '../../../components/symbol/symbolsComponent'
+import Symbols from '../../../components/symbol/index/index'
 
 class Symbol extends React.Component {
   constructor(props) {
@@ -11,7 +11,7 @@ class Symbol extends React.Component {
 
   toggleSymbolForm = () => {
     this.setState({showSymbolForm: !this.state.showSymbolForm})
-  } 
+  }
 
   render() {
     return <div className="">
